check-universe command display available trading pairs
